+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -19,7 +19,8 @@ exclude: true
   ([bug 1618058]({{bugzilla}}1619766))
 - Added [`WebNotification.source`][79.2], which is the URL of the page
   or Service Worker that created the notification.
-
+- Removed deprecated `WebExtensionController.setTabDelegate` and `WebExtensionController.getTabDelegate`
+  APIs ([bug 1618987]({{bugzilla}}1618987)).
 
 [79.1]: {{javadoc_uri}}/WebExtension.TabDelegate.html#onOpenOptionsPage-org.mozilla.geckoview.WebExtension-
 [79.2]: {{javadoc_uri}}/WebNotification.html#source
@@ -723,4 +724,4 @@ exclude: true
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 820c4bbc052bd8b8b19c2460c3996b86e1786ff7
+[api-version]: f395ea7c21fe233e4e95d28f255910bcde75dc91